requirements.txt
asgiref==3.9.1
Django==5.2.5
pillow==11.3.0
sqlparse==0.5.3
python-dotenv==1.1.1
psycopg2-binary==2.9.10
djangorestframework==3.15.2
channels==4.2.0
channels_redis==4.2.1
djangorestframework-simplejwt==5.5.0
daphne==4.2.0